Factor isecho into platform-specifics
