+++ cmd/win/win_windows.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"os"
+)
+
+func isecho(fp *os.File) bool {
+	return true
+}
